backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.35
alembic==1.15.1
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
celery[redis]==5.4.0
redis==5.2.1
cryptography==43.0.1
python-jose[cryptography]==3.4.0
python-multipart==0.0.22
weasyprint==68.0
reportlab==4.2.5
pandas==2.2.3
openpyxl==3.1.5
httpx==0.27.1
supabase==2.5.1
unstructured[pdf]==0.16.0
pytest==8.3.3
pytest-asyncio==0.24.0
aiosqlite==0.20.0